imdbproj/requirements.txt
mysql-connector-python
sqlalchemy
pandas
matplotlib
seaborn
